Changed return value of State::getValue() into size_t in the header. This was unsigned int, being a discrepancy between implementation and declaration. This was reported by Bas Ploeger. He also reported some warnings caused by graveyard code, and an error for systems in which DISPLAY does not have any value. I will fix this next week.
git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@4875 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ tools/gltsgraph/state.h
@@ -36,7 +36,7 @@ class State
     double getY() const;
 
     std::string getLabel() const;
-    unsigned int getValue() const;
+    size_t getValue() const;
 
     size_t getNumberOfTransitions() const;
     size_t getNumberOfInTransitions() const;
